refactor: remote styledcomponent in space component and to set up type of image and rules of eslint
