+++ .eslintrc.json
@@ -32,6 +32,7 @@
     "no-unused-vars": "off", //타입스크립트 사용시 interface의 변수명을 eslint가 잡지 않도록 함.
     "@typescript-eslint/no-unused-vars": "warn", //대신 사용하지 않는 변수는 @typescript/eslint를 통해 잡아줌.
     "simple-import-sort/imports": "error", // eslint-plugin-simple-import-sort 설정
-    "simple-import-sort/exports": "error" // eslint-plugin-simple-import-sort 설정
+    "simple-import-sort/exports": "error", // eslint-plugin-simple-import-sort 설정
+    "react/require-default-props": "off"  //component prop 타입체크 끄기
   }
 }
