+++ package.json
@@ -14,7 +14,6 @@
     "@vercel/analytics": "^1.1.1",
     "axios": "^1.5.1",
     "dotenv": "^16.3.1",
-    "eslint-plugin-simple-import-sort": "^10.0.0",
     "ioredis": "^5.3.2",
     "memory-cache": "^0.2.0",
     "next": "13.5.4",
@@ -37,6 +36,7 @@
     "autoprefixer": "^10",
     "eslint": "^8.2.0",
     "eslint-config-airbnb": "^19.0.4",
+    "eslint-plugin-simple-import-sort": "^10.0.0",
     "eslint-config-airbnb-typescript": "^17.1.0",
     "eslint-config-next": "13.5.4",
     "eslint-config-prettier": "^9.0.0",
