+++ src/model/image.d.ts
@@ -1,0 +1,4 @@
+// src/types/images.d.ts
+declare module '*.png';
+declare module '*.jpg';
+declare module '*.jpeg';
