+++ tsconfig.json
@@ -22,6 +22,7 @@
       "@/*": ["./src/*"]
     }
   },
+  "typeRoots": ["src/model"],
   "include": ["next-env.d.ts", "**/*.ts", "**/*.tsx", ".next/types/**/*.ts"],
   "exclude": ["node_modules"]
 }
